+++ src/components/Skeleton/BoxSkeleton.tsx
