+++ src/components/index.ts
@@ -1,40 +1,40 @@
-export { default as Accordion } from './_common/Accordion/Accordion';
-export { default as Button } from './_common/Button/Button';
-export { default as LineChart } from './_common/Chart/LineChart';
-export { default as RadarChart } from './_common/Chart/RadarChart';
-export { default as Drawer } from './_common/Drawer/Drawer';
-export { default as Dropdown } from './_common/Dropdown/Dropdown';
-export { default as AuthErrorFallback } from './_common/Error/AuthErrorFallback';
-export { default as CountBoxFallback } from './_common/Error/CountBoxFallback';
-export { default as ErrorBox } from './_common/Error/ErrorBox';
-export { default as ErrorResetBoundary } from './_common/Error/ErrorResetBoundary';
-export { default as GlobalErrorBoundary } from './_common/Error/GlobalErrorBoundary';
-export { default as GlobalErrorFallbackSection } from './_common/Error/GlobalErrorFallbackSection';
-export { default as GlobalSuspense } from './_common/Error/GlobalSuspense';
-export { default as PageErrorFallback } from './_common/Error/PageErrorFallback';
-export { default as SectionErrorFallback } from './_common/Error/SectionErrorFallback';
-export { default as Flex } from './_common/Flex/Flex';
-export { default as FormField } from './_common/FormField/FormField';
-export { default as Header } from './_common/Header/Header';
-export { default as HeaderMobile } from './_common/Header/HeaderMobile';
-export { default as Heading } from './_common/Heading/Heading';
-export { default as Input } from './_common/Input/Input';
-export { default as AuthGuard } from './_common/Layout/AuthGuard';
-export { default as DrawerLayout } from './_common/Layout/DrawerLayout';
-export { default as Layout } from './_common/Layout/Layout';
-export { default as Main } from './_common/Main/Main';
-export { default as AlertModal } from './_common/Modal/AlertModal';
-export { default as Modal } from './_common/Modal/Modal';
-export { default as NavigationBar } from './_common/NavigationBar/NavigationBar';
-export { default as BoxSkeleton } from './_common/Skeleton/BoxSkeleton';
-export { default as TableListSkeleton } from './_common/Skeleton/TableListSkeleton';
-export { default as TableSkeleton } from './_common/Skeleton/TableSkeleton';
-export { default as Table } from './_common/Table/Table';
-export { default as Tabs } from './_common/Tabs/Tabs';
-export { default as Text } from './_common/Text/Text';
-export { default as Title } from './_common/Title/Title';
-export { default as ToggleButton } from './_common/ToggleButton/ToggleButton';
-export { default as UploadButton } from './_common/UploadButton/UploadButton';
+export { default as Accordion } from './Accordion/Accordion';
+export { default as Button } from './Button/Button';
+export { default as LineChart } from './Chart/LineChart';
+export { default as RadarChart } from './Chart/RadarChart';
+export { default as Drawer } from './Drawer/Drawer';
+export { default as Dropdown } from './Dropdown/Dropdown';
+export { default as AuthErrorFallback } from './Error/AuthErrorFallback';
+export { default as CountBoxFallback } from './Error/CountBoxFallback';
+export { default as ErrorBox } from './Error/ErrorBox';
+export { default as ErrorResetBoundary } from './Error/ErrorResetBoundary';
+export { default as GlobalErrorBoundary } from './Error/GlobalErrorBoundary';
+export { default as GlobalErrorFallbackSection } from './Error/GlobalErrorFallbackSection';
+export { default as GlobalSuspense } from './Error/GlobalSuspense';
+export { default as PageErrorFallback } from './Error/PageErrorFallback';
+export { default as SectionErrorFallback } from './Error/SectionErrorFallback';
+export { default as Flex } from './Flex/Flex';
+export { default as FormField } from './FormField/FormField';
+export { default as Header } from './Header/Header';
+export { default as HeaderMobile } from './Header/HeaderMobile';
+export { default as Heading } from './Heading/Heading';
+export { default as Input } from './Input/Input';
+export { default as AuthGuard } from './Layout/AuthGuard';
+export { default as DrawerLayout } from './Layout/DrawerLayout';
+export { default as Layout } from './Layout/Layout';
+export { default as Main } from './Main/Main';
+export { default as AlertModal } from './Modal/AlertModal';
+export { default as Modal } from './Modal/Modal';
+export { default as NavigationBar } from './NavigationBar/NavigationBar';
+export { default as BoxSkeleton } from './Skeleton/BoxSkeleton';
+export { default as TableListSkeleton } from './Skeleton/TableListSkeleton';
+export { default as TableSkeleton } from './Skeleton/TableSkeleton';
+export { default as Table } from './Table/Table';
+export { default as Tabs } from './Tabs/Tabs';
+export { default as Text } from './Text/Text';
+export { default as Title } from './Title/Title';
+export { default as ToggleButton } from './ToggleButton/ToggleButton';
+export { default as UploadButton } from './UploadButton/UploadButton';
 
-export * from './_common/Deferred/DeferredComponent';
-export * from './_common/Spinner/Spinner';
+export * from './Deferred/DeferredComponent';
+export * from './Spinner/Spinner';
